+++ RELEASES.md
@@ -1099,3 +1099,13 @@ Based on:
 - [typescript v0.6.0-alpha.17] .
 ### Releases
 - [NPM v0.6.0-alpha.17] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.17 - .
+
+## 2026-02-21 02:09:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.723.0 (2.835.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0-alpha.18] .
+### Releases
+- [NPM v0.6.0-alpha.18] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.18 - .